remove CardsList component + move ApplicationCard Component
