--- apps/chat/src/components/Marketplace/CardsList.tsx
@@ -1,58 +1,0 @@
-import { memo } from 'react';
-
-import { DialAIEntityModel } from '@/src/types/models';
-
-import { ApplicationCard } from '@/src/components/Marketplace/ApplicationCard';
-
-import { PublishActions } from '@epam/ai-dial-shared';
-
-interface CardsListProps {
-  entities: DialAIEntityModel[];
-  className?: string;
-  onCardClick: (entity: DialAIEntityModel) => void;
-  onPublish?: (entity: DialAIEntityModel, action: PublishActions) => void;
-  onDelete?: (entity: DialAIEntityModel) => void;
-  onEdit?: (entity: DialAIEntityModel) => void;
-  onBookmarkClick?: (entity: DialAIEntityModel) => void;
-  onSelectVersion?: (entity: DialAIEntityModel) => void;
-  onLogsClick?: (entity: DialAIEntityModel) => void;
-  dataQA?: string;
-}
-
-export const CardsList = memo(
-  ({
-    entities,
-    className,
-    onCardClick,
-    onPublish,
-    onDelete,
-    onEdit,
-    onBookmarkClick,
-    onLogsClick,
-    dataQA,
-  }: CardsListProps) => {
-    return (
-      <section className={className}>
-        <div
-          className="grid grid-cols-1 gap-3 md:grid-cols-2 md:gap-4 xl:grid-cols-3 xl:gap-5 3xl:grid-cols-4 4xl:grid-cols-5 5xl:grid-cols-6"
-          data-qa={dataQA}
-        >
-          {entities.map((entity) => (
-            <ApplicationCard
-              key={entity.id}
-              entity={entity}
-              onPublish={onPublish}
-              onDelete={onDelete}
-              onClick={onCardClick}
-              onEdit={onEdit}
-              onLogsClick={onLogsClick}
-              onBookmarkClick={onBookmarkClick}
-            />
-          ))}
-        </div>
-      </section>
-    );
-  },
-);
-
-CardsList.displayName = 'CardsList';
